moved parsing of the data around
